Refactored reader_setter.
- It now supports a return pointer to the data node with the file.
- Changed some calls to it to accept this data node, so that log_error can report the node if something goes wrong.
- General refactor to mob type property loading code, by sorting them alphabetically and making them more consistent.

+++ Source/source/mob_types/pellet_type.h
@@ -25,11 +25,11 @@ class pellet_type : public mob_type {
 public:
     pikmin_type* pik_type;
     //Number on the pellet, and hence, its weight.
-    unsigned number;
+    size_t number;
     //Number of seeds given out if the pellet's taken to a matching Onion.
-    unsigned match_seeds;
+    size_t match_seeds;
     //Number of seeds given out if the pellet's taken to a non-matching Onion.
-    unsigned non_match_seeds;
+    size_t non_match_seeds;
     ALLEGRO_BITMAP* bmp_number;
     
     pellet_type();
